Fix a bug that causes driftdetector generates a wrong message
ResourceKey has a namespace field. For live objects that field is not empty but most of the git objects they are emtpy.
This causes comparing the resource key between live object and git object will be not correct.
So we add another comparation function with ignoring the namespace to use.