Fix a bug that causes driftdetector generates a wrong message (#132)
**What this PR does / why we need it**:

ResourceKey has a namespace field. For live objects, that field is not empty but most of the git objects they are emtpy.
This causes comparing the resource key between live object and git object will be not correct.
So we add another comparation function with ignoring the namespace to use.

**Which issue(s) this PR fixes**:

Fixes #

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.