Append dummy path to file creation URL (#637)
**What this PR does / why we need it**:
We're getting an issue with specifying a filename: https://github.com/isaacs/github/issues/1527
In short, the last path part of the URL is ignored when specifying a filename. For now, appending dummy path as a workaround.

**Which issue(s) this PR fixes**:

Fixes #

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.