Simplify piped controller by introducing a new deploysource package (#1101)
**What this PR does / why we need it**:

Sorry for this big PR. I was unable to break it into smaller ones.
This PR introduces a new `deploysource` package and applies it in some packages for piped such as scheduler, planner, executor.
By using this new package, we can archive the following benefits:
- decoupling the source/config preparing for the running commit and the target commit
(Because the rollback stage does not require the source/config of the target commit. So by decoupling them, when the target commit/configuration could not be loaded successfully, the rollback stage will not be affected)
- preparing source will be lazied. This means they will be loaded only when needed
- DRY

**Which issue(s) this PR fixes**:

Fixes #

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.