Add image provider for ECR (#1213)
**What this PR does / why we need it**:
This PR does:
- add ECR client to fetch the latest tag
- add `semver` package to parse and sort tags that is according to Semantic Versiong.

This mainly aims to let `GetLatestImage` work properly to determine the latest tag. Hopefully AWS provides the API to give back the latest tag in the specified repository. But real-world is not so easy. As far as I investigate, the only way to determine the latest tag is:

- Fetch all tags by issuing the request to ListImages API.
- Then fetch the time pushed at for each tag by calling DescribeImages API.

To avoid reaching the API rate limit, it attempts to determine by the semantic versioning as much as possible.

The way to list all tags doesn't scale, so I left a TODO comment and will be a future issue.

**Which issue(s) this PR fixes**:

Fixes #

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.