Fix no namespace resource found on delete resource k8s (#1281)
**What this PR does / why we need it**:

The `namespace` in KubernetesDeploymentInput config currently is not a required field. In case of delete resource on unset that namespace field, namespace default will be used ( as a default namespace of k8s ) leads to `not found resources to delete` bug.

**Which issue(s) this PR fixes**:

Fixes https://github.com/pipe-cd/pipe/issues/1275

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.