Add notes on copying git repository (#1296)
**What this PR does / why we need it**:
On MacOS, BSD's `cp` command is installed by default. So the behavior is slightly different from GNU's one.
This PR will cause the build of the binary for Windows to fail. I don't mind add `command_windows.go` if it seems to be a problem.

**Which issue(s) this PR fixes**:

Fixes #

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.