Implement Event watcher (#1411)
**What this PR does / why we need it**:
Add Event watcher, a piped component that periodically fetches the latest Event from the `eventstore` to compare the value with one in the git repository.

Most core logic is borrowed from `imagewatcher` being already. And soon-to-be enables this as another PR.

**Which issue(s) this PR fixes**:

Ref https://github.com/pipe-cd/pipe/issues/1381

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.