Move role config to Function manifest instead of Piped manifest (#1439)
**What this PR does / why we need it**:

The role parameter passed to Create/Update Lambda application is the role set for the deploying lambda function (service role), not for the running agent (user role) to deploy it. So, I think it should be a part of FunctionManifest instead of CloudProviderConfig as the current design.

**Which issue(s) this PR fixes**:

Fixes #

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.