Fix the diffs on Istio's VirtualService always appear after deploying (#1504)
**What this PR does / why we need it**:

With the current implementation of K8S_TRAFFIC_ROUTING stage, it is also appending the zero-weight destinations into the VirtualService manifest. That causes the diffs after deploying when those destinations were not specified by users.

This pull request switches to directly use the manifest defined in the target commit when the primary weight is 100% to avoid those diffs.

**Which issue(s) this PR fixes**:

Fixes #1492

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.