Remove lock file for dependencies of pipecd chart (#1876)
**What this PR does / why we need it**:
This PR makes the process to update dependency charts simpler.
In CI step for publishing charts, it runs `helm package` with `--dependency-update`, which automatically generates Chart.lock. Because of that, the only thing that we need to do when updating verion is modifying Chart.yaml.

**Which issue(s) this PR fixes**:

Fixes https://github.com/pipe-cd/pipe/issues/1863

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.