Enable drift detector to compare number and numeric string (#1875)
**What this PR does / why we need it**:

Some Kubernetes resource fields are using `resources.Quality` data type (a customized struct to allow specifying cpu/memory by either number or string), that causes inconsistent marshaling result with the defined one in Git.

Adding edge-handling for the case such as `resource.cpu`, `resource.memory` is not enough because so many CRDs are also using that data type too, not just standard resource like Deployment.

So I think we can enable our drift detector component to compare number values and numeric strings to avoid those diffs.
The downside of this way is that other fields (not `resource.cpu` or `resource.memory`) also will be the target of this comparation.
But I think once they are the wrong type, they will be failed while applying to the Kubernetes cluster. So the drift detector can rely on that assumption.)

**Which issue(s) this PR fixes**:

Fixes #

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
Enable drift detector to compare number and numeric string
```

This PR was merged by Kapetanios.