Update piped-base to not depend on /home/pipecd (#1871)
**What this PR does / why we need it**:
Until now, we assumed the home directory of the user Piped runs as is `/home/pipecd`. But we can't specify the user name at the pod template. So I settled on going back to use `/etc/ssh/ssh_config` as an ssh config file and `/tools` as a binary directory. Let me check how it goes once got merged because I have no chance to build the container locally.

**Which issue(s) this PR fixes**:

Fixes 

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.