Honor $HOME so that any user can write things in a unified way (#1899)
**What this PR does / why we need it**:
We settled on putting all files to be touched by Piped into the home directory. The piped-base image is already changed to use them: https://github.com/pipe-cd/pipe/pull/1890

I know this PR includes a known issue. It can happen only if the Piped process was owned by a kind of user not in `/etc/passwd`. I just settled on putting this issue to https://github.com/pipe-cd/pipe/issues/1905. And I'd love to merge this PR for now because the latest version of piped doesn't work even if it's running by a user existing in `/ect/passwd`. We had better fix it as soon as possible.

**Which issue(s) this PR fixes**:

Fixes https://github.com/pipe-cd/pipe/issues/1885

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.