Support contains operator for MySQL (#2042)
**What this PR does / why we need it**:
Indexing on a JSON column didn't go well. Though I'm in the middle of trying to do it, the query works fine without the index.
Let me merge it to support the environment deletion feature in the next version. It's no problem to add index after that version.

**Which issue(s) this PR fixes**:

Fixes https://github.com/pipe-cd/pipe/issues/2033

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.