Disable EnvIds column (#2049)
**What this PR does / why we need it**:
In order to maintain consistency, commented out the newly added column where type is likely to be varied in the near future. After fixing the type, it will be reverted.

If you try to delete an environment where datastore is MySQL, it gets something like:

![image](https://user-images.githubusercontent.com/19730728/121148135-733fe800-c87c-11eb-8eee-a4f2be111ea0.png)

**Which issue(s) this PR fixes**:

Fixes #

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.