Add ability to use both SecretManagement and SealedSecretManagement (#2084)
**What this PR does / why we need it**:

This PR adds a new `SecretManagement` field next to the existing `SealedSecretManagement` in `Piped` configuration. The new one is more general so it can be used widely, from the current feature as `SealedSecret` to the future ones as  `Secret Templating`, `Remote Config`.
To ensure backward compatibility the`SealedSecretManagement` field is marked with the "Deprecated" label and will be kept for a while. It will be removed completely in the future.

**Which issue(s) this PR fixes**:

Fixes #

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.