Fix: obsolete terraform flags (#2138)
**What this PR does / why we need it**:

Terraform cloudprovider failed to execute with the latest terraform stable version 1.0.0. This PR addresses the issues caused by changes to command line flags, and 

The directory parameter '.' has been changed to `-chdir` and is optional. This PR removes the flag from the command
The `-lock=false` is no longer valid parameter for `init` command. This PR removes the flag from the command. 
I've tested the changed parameters with following terraform versions:
- 1.0.0
- 0.15.5
- 0.14.11
- 0.13.7

Parsing plan with no changes failed in latest terraform version. This is most likely due to subtle changes to the text output. 

![image](https://user-images.githubusercontent.com/12850042/123560326-6cc8d000-d799-11eb-828d-8dfbb80cebc4.png)


This PR proposes change to detect explicit exit status to parse only plans with changes. This method should be less error prone when the string output of terraform changes. 

**Which issue(s) this PR fixes**:


**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.