Change to manage Grafana datasources as a ConfigMap (#2227)
**What this PR does / why we need it**:
We used to fix the Prometheus service name as `pipecd-prometheus-server` to discover it from Grafana statically. This PR resolves that problem by changing to manage the datasources config as our ConfigMap, which can template with the release name.

On the other hand, still we have to fix that ConfigMap name. It is `pipecd-grafana-datasources` by default. If you want to change the name, you can change it by setting `.Values.grafana.extraConfigmapMounts[0].configMap`.

**Which issue(s) this PR fixes**:

Fixes https://github.com/pipe-cd/pipe/issues/2226

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.