Add relabel configs to identify gRPC service name (#2250)
**What this PR does / why we need it**:
This PR adds the Prometheus relabel configs to extract gRPC service names to be used for Grafana variables (candidates are `piped`, `web` and `api`).

![image](https://user-images.githubusercontent.com/19730728/125599737-5f10a0be-b3fa-47c1-9940-3cf7e26008fc.png)

It also changes the Envoy cluster names in order to easily relabel using regex. Tell me a better name if any (actually whatever is fine as long as it's according to some kind of pattern)


**Which issue(s) this PR fixes**:

Fixes #

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.