Fix datasource not found for provisioned dashboards (#2469)
**What this PR does / why we need it**:

While initializing provisioned dashboards from exported JSON configurations, all fields such as `"__input"` which contains defined variable `$DS_PROMETHEUS` is omitted that cause `$DS_PROMETHEUS datasource not found` errors. This should be counted as a Grafana issue but currently, patches for this are not available and we have this workaround (addressed by this PR) for this.

ref: https://github.com/grafana/grafana/issues/10786

**Which issue(s) this PR fixes**:

Fixes #

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.