API add ability to filter deployments by application name (#2493)
**What this PR does / why we need it**:

Currently, we only have a way to filter deployments by Env, Kind, Status, and ApplicationId (display as application name on UI) which is hard to use in case there are many applications with the same name. This change adds ApplicationName as a filter, the change for the client will be addressed by another PR.

**Which issue(s) this PR fixes**:

Part of #2448

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.