Use metrics analyzer to perform ADA for each strategy (#2531)
**What this PR does / why we need it**:
This PR replaces the analyzer with the new analyzer that can perform ADA in the four strategies. Except for `THRESHOLD`, implementation is not complete yet, but it's fine because we don't publish to users the new configuration on the docs.

In short, this doesn't change any behavior. Just preparing for the new strategies.

**Which issue(s) this PR fixes**:

Fixes https://github.com/pipe-cd/pipe/issues/2506

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.