Introduce launcher a component to manage Piped agent (#2557)
**What this PR does / why we need it**:

I keep Launcher stays inside Piped container image and make it as a sub-command of Piped binary for this time.
And I think it is still good enough and makes things simple.
Because currently, we are already having 2 images for Piped (general one and OKD one). If Launcher is built as a separate container image, we will have 4 Piped images in all.

**Which issue(s) this PR fixes**:

Fixes #

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
Introduce launcher a component to manage Piped agent
```

This PR was merged by Kapetanios.