Empower all commands to handle input to stdin (#2598)
**What this PR does / why we need it**:
This PR makes it possible to pass the input to stdin to all commands.
Due to this, it renamed `cli.Telemetry` to `cli.Input`. If you think about it, originally logger isn't in the telemetry context so not odd to change. This brought flexibility to us for a variety of other inputs to be propagated to each command.

The initial reason why I looked to make this change is so that we can pass the plaintext file to `pipectl encrypt` command via stdin.

**Which issue(s) this PR fixes**:

Fixes #

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.