Fix bug always use the same livestate store in case of multiple cloudproviders of the same kind are set (#2638)
**What this PR does / why we need it**:

Currently, in case of multiple cloud-providers of the same kind are set in Piped configuration, only one store is used to handle the livestate's works. Even we have a map that contains multiple livestate stores pointed by the cloud-provider's name, we always pass the same key to get the last livestate store from the map.
Ref: https://golang.org/doc/faq#closures_and_goroutines

**Which issue(s) this PR fixes**:

Fixes #2622

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
Fix bug always use the same livestate store in case of multiple cloud-providers of the same kind are set
```

This PR was merged by Kapetanios.