Add an admin HTTP server to launcher (#2667)
**What this PR does / why we need it**:

In order to run Launcher as a CloudRun service, it is required to have at least one HTTP server to serve incoming requests.
I tried to use the `admin` server of its Piped for that requirement but CloudRun is killing the Launcher container immediately because launching Piped and its server takes time. Since CloudRun is not providing any way to delay that readiness probe currently, so this PR adds another `admin` HTTP server into Launcher.

**Which issue(s) this PR fixes**:

Fixes #

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.