Add project info to notification events' messages (#2679)
**What this PR does / why we need it**:

In case of working on multiple projects (just as us on the PipeCD development environment), this project information would help. Also, we can use it as a query value for the link to the PipeCD console with `?project=` key to ensure redirection to the right project for the notified events.

**Which issue(s) this PR fixes**:

Fixes #

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
Add Project information to notification events' messages
```

This PR was merged by Kapetanios.