Fix wrong namespace value shown in deployment log (#2681)
**What this PR does / why we need it**:

In case the namespace for Kubernetes app is specified in PipeCD deployment configuration file (path `spec.input.namespace`), the wrong namespace value ("default" or ns value in its manifests) will be shown in the deployment log.

Basically, the value in PipeCD deployment configuration file has a higher priority and will be used on applying manifest to the cluster. This change omits that wrong value from deployment log to avoid confusion on the deployment log.

**Which issue(s) this PR fixes**:

Fixes #2637 

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.