Add sync stable docs command (#2724)
**What this PR does / why we need it**:

To reduce manual steps while updating docs of the latest released version, lets add a command to sync the change on the latest version docs to the stable docs. A typical use-case is: Make changes on specified released version docs (usually the latest version), run command to sync all the changes of that version docs as stable latest docs.

The command is: `make sync-stable-docs version=v0.21.x`

**Which issue(s) this PR fixes**:

Fixes #

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.