Move TriggeredDeploymentID metadata key to command model (#2800)
**What this PR does / why we need it**:

`triggeredDeploymentIDKey` used in both trigger(piped) and pipectl but must be defined to use could lead to unexpected error in case the key name changed in one end. This PR moves the key to the Command model (since it is the key in command metadata) and explicitly refers it in the used end.

**Which issue(s) this PR fixes**:

Fixes #

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.