Add an rpc to update registered apps that have been changed (#2830)
**What this PR does / why we need it**:
This PR adds an RPC to update registered enabled apps.
As discussed previously, it is based on the assumption that a single app config file will not be referenced by multiple enabled apps. This should be made explicit in the documentation. But well eventually this will be able to be validated when added.

One case where is likely happen is enabling a disabled app that refers the same app config file. Possibly this can be addressed by the validation.

**Which issue(s) this PR fixes**:

Fixes https://github.com/pipe-cd/pipe/issues/2803

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.