Add status and reason to deployment ref in deployment chain model (#2864)
**What this PR does / why we need it**:

This PR adds deployment's status and status reason to the deployment ref in the deployment chain model. It will reduce the datastore read in case we want to use those values to determine deployment chain behaviours, besides, we will use those values to show in the client too.

**Which issue(s) this PR fixes**:

Fixes #2836

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.