Cancel deployments of the same block in case the block finished with FAILURE or CANCELLED status (#2893)
**What this PR does / why we need it**:

In case some running deployments are in the same block with a FAILURE or CANCELLED deployment, we should stop those deployments too. By this PR, I added a method to check whether a deployment should be stopped or not when its piped sync the on processing deployment status via `ReportDeploymentPlanned` and `ReportDeploymentStatusChanged`, at that point of time, deployment is still on processing and if the current block which contains that deployment finished with FAILURE | CANCELLED status, we will stop the current deployment too by sending CANCEL_DEPLOYMENT command so that the handling piped will be aware and stop its deployment.

Note: I don't use the same method (add `cancel` flag to those gRPC responses) as we do with `InChainDeploymentPlannable` since
- This time, the gRPC which be updated is called once, not intervally as check plannable RPC.
- I want to keep the deployment handling mechanism on piped side simple (reuse the cancelling by command flow in planner and scheduler codebase)

**Which issue(s) this PR fixes**:

Fixes #2888 

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.