Add a form to add an app from Git (#2894)
**What this PR does / why we need it**:
This PR adds React components to select unregistered applications.
To avoid bloating the PR size, let me separate the API part into another PR.

The newly added components are `UnregisteredApplicationList` and `UnregisteredApplicationFilter`. I would be happy to get feedback from you on the code as well as the UI in detail.

Note that this GIF animation is based on temporary hard-coded dummy data, so you can't reproduce it.

![Kapture 2021-12-07 at 16 46 23](https://user-images.githubusercontent.com/19730728/144987707-c48d1563-93a5-4c2d-a6cb-bbe6b0a6510a.gif)

**Which issue(s) this PR fixes**:

Ref https://github.com/pipe-cd/pipe/issues/2885

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.