Make it possible to register labels via app config (#2963)
**What this PR does / why we need it**:
With it, you will be able to register labels defined in your application configuration file but this is still not documented. It just aims to make debugging easier.

**Which issue(s) this PR fixes**:

Fixes https://github.com/pipe-cd/pipe/issues/2913

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.