Change default interval to scan app config files (#2966)
**What this PR does / why we need it**:
In case that you have so many unregistered application config files under your repository, it might be not fine to scan files per minute. But it hardly happens. Typically, you would be happy if the application configuration file is reflected in the web console as soon as possible after you add it.

**Which issue(s) this PR fixes**:

Fixes #

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.