Fix event watcher not to handle already handled events (#3108)
**What this PR does / why we need it**:
Even though the status of Event was updated, the status of the latest event cached by Piped had not changed. Hence we've been facing an issue that Piped tries to handle handled Events unless we restarted Piped.

This PR addresses it.

**Which issue(s) this PR fixes**:

Fixes #

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.