Handle unavalable piped to avoid blocking plan-preview (#3181)
**What this PR does / why we need it**:
Currently, when piped does not execute command because it is offline or for some other reason, plan-preview requests are blocked and results in a timeout. To avoid this, if piped is not online, return the result without waiting command execution.

**Which issue(s) this PR fixes**:

Fixes #3102

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.