Drop piped table env_ids index (#3206)
**What this PR does / why we need it**:

This PR removes the unneeded index `piped_project_id_env_ids_asc` on Piped table. The queries on MySQL datastore work with/without index but we chose to make that index since this `EnvIds` field is a virtual generated column, which its value will be generated every time we query for it, which makes the operation a bit slower.
Since the `Environment` concept is deprecated and will be removed in the near future, I think it's fine to just remove this index to get rid of environment-related unstable queries issues. Besides, the number of piped(s) which be controlled by a control-plane is not too many, query without index is acceptable at this meantime.

Tested on my local, some places we needed to use this index previously

![Screen Shot 2022-02-06 at 14 20 25](https://user-images.githubusercontent.com/32532742/152668706-9527fa4e-9be3-4122-b0f9-fcc76d8c8c9f.png)

![Screen Shot 2022-02-06 at 14 21 17](https://user-images.githubusercontent.com/32532742/152668722-7e75103d-8a21-4a4f-b259-3fce139213ad.png)

**Which issue(s) this PR fixes**:

Fixes #3191

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.