Introduce commander to models store interface (#3243)
**What this PR does / why we need it**:

The idea has remained from PR #3144 but I change the implementation based on the comment https://github.com/pipe-cd/pipecd/pull/3144#discussion_r794164550

This PR introduces a new field named `writer` to each models' store, the `store.writer` will be used by Update-related functions or whenever we need to know who is the caller of that datastore operation.
Currently, we have 4 known writers which are: web, pipectl, ops, piped. In this PR, I added 2 more writers which are: TestWriter and UnknownWriter. The first is dedicated to testing, the latter is used in case there is no way to figure out who will be the caller for that datastore operation (there is one like that, it's used by HTTP API server in cmd/pipecd/server.go).

Note for the UnknownWriter, as the next work after this PR is merged, I planned to make logic like UnkownWriter is unable to order write operation to the datastore or such in datastore layer, please lets me know if you guys have a better idea to handle that kind of thing.

> Update: Writer is renamed as Commander in the latest implementation

**Which issue(s) this PR fixes**:

Fixes #

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.