Move architecture overview image to concept explanation page (#2994)
**What this PR does / why we need it**:

In the current concept explanation page, we have definitions of PipeCD's components such as: piped, control-plane, etc. but do not have an image to explain what it is in a big picture so it's hard to follow, the image which shows PipeCD architecture overview is on the overview page, which contains stub not really related to the thing shown in the image. So my idea is to move the architecture overview image to the concepts explanation page and use a different image to show what PipeCD is on the overview page as below:

Overview:
![Screen Shot 2022-02-21 at 12 24 51](https://user-images.githubusercontent.com/32532742/154884136-cffacfd0-d7d8-410c-a47b-e7a7c57c3da9.png)

Concepts:
<img width="948" alt="Screen Shot 2021-12-29 at 22 29 15" src="https://user-images.githubusercontent.com/32532742/147667286-2d2c83c0-af11-4647-8987-16c5c7d56a43.png">

**Which issue(s) this PR fixes**:

Fixes #

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.