Add merge shards data for multiple shards stored object (#3316)
**What this PR does / why we need it**:

This PR introduces a way to merge raw data for multiple shards stored objects. I defined a new interface `ShardDecoder`, which will be used in case of building object from fetched data in filedb.

The PR also contains the implementation of the `ShardDecoder` interface for only `commandCollection` since the Command model is the only model which is multiple shards stored and there is a collision on its field named `Status` (should be stored in 2 shards: agent and piped)

**Which issue(s) this PR fixes**:

Fixes #

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.