Add etag to the object attribute struct (#3335)
**What this PR does / why we need it**:

We need this attribute to detect whether the object content is updated or not. It will be used to reduce the number of unnecessary fetch file content requests to the filestore service.

**Which issue(s) this PR fixes**:

Fixes #

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.