Add index to CompletedAt (#3388)
**What this PR does / why we need it**:
Currently, completedAt does not exists on mysql column, and so order by completedAt will be failed.
To fix this, this PR create generated column of completedAt and add index to it. CompletedAt is nullable.

**Which issue(s) this PR fixes**:

Fixes #

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.