Add filedb filter logic (#3389)
**What this PR does / why we need it**:

This PR adds filter logic which is used in the Find interface. Previously, I planned to make filter logic as part of collection logic and implement it for each model kind. Since the filter logic accepts `Filter.Field` (a string) as its input. If we make that filter logic as part of collections, we may have to use reflect to list up struct's fields name as string, it's complicated and not worth it. In this PR, instead of making filter logic as part of collections, it will be addressed in filedb layer, with a more generic interface `map[string]interface{}` as the input entity.

We still accept filter logic in collections (if exists) and only use this generic filter logic in case the collection does not implement the `filterable` interface.

**Which issue(s) this PR fixes**:

Fixes #

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.