Fix at the beginning of the sentence for shortening terraform plan details (#3467)
**What this PR does / why we need it**:
Fix the rule of shortening terraform plan details in actions-plan-preview because the sentence is different for each version of Terraform as the following.


**v0.12.26**
```
An execution plan has been generated and is shown below.
Resource actions are indicated with the following symbols:
  + create

Terraform will perform the following actions:
```

**v1.0.3**
```
Terraform used the selected providers to generate the following execution
plan. Resource actions are indicated with the following symbols:
  - destroy

Terraform will perform the following actions:
```

**Which issue(s) this PR fixes**:

Fixes #

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
-->
```release-note
NONE
```

This PR was merged by Kapetanios.